+++ doc/README_Sink.md
@@ -94,7 +94,7 @@ Refer to the [sink properties](#sink-configuration-properties) section for more 
 {
   "name": "cosmosdb-sink-connector",
   "config": {
-    "connector.class": "com.microsoft.azure.cosmosdb.kafka.connect.sink.CosmosDBSinkConnector",
+    "connector.class": "com.azure.cosmos.kafka.connect.sink.CosmosDBSinkConnector",
     "tasks.max": "1",
     "topics": [
       "hotels"
@@ -167,7 +167,7 @@ The following settings are used to configure the Cosmos DB Kafka Sink Connector.
 | Name | Type | Description | Required/Optional |
 | :--- | :--- | :--- | :--- |
 | topics | list | A list of Kafka topics to watch | Required |
-| connector.class | string | Classname of the Cosmos DB sink. Should be set to `com.microsoft.azure.cosmosdb.kafka.connect.sink.CosmosDBSinkConnector` | Required |
+| connector.class | string | Classname of the Cosmos DB sink. Should be set to `com.azure.cosmos.kafka.connect.sink.CosmosDBSinkConnector` | Required |
 | connect.cosmosdb.connection.endpoint | uri | Cosmos DB endpoint URI string | Required |
 | connect.cosmosdb.master.key | string | The Cosmos DB primary key that the sink connects with | Required |
 | connect.cosmosdb.databasename | string | The name of the Cosmos DB database the sink writes to | Required |
