+++ src/docker/resources/sink.config.json
@@ -1,7 +1,7 @@
 {
   "name": "cosmosdb-sink-connector",
   "config": {
-    "connector.class": "com.microsoft.azure.cosmosdb.kafka.connect.sink.CosmosDBSinkConnector",
+    "connector.class": "com.azure.cosmos.kafka.connect.sink.CosmosDBSinkConnector",
     "tasks.max": "1",
     "topics": "connect-test",
     "value.converter": "org.apache.kafka.connect.json.JsonConverter",
