+++ src/main/resources/log4j.properties
@@ -6,4 +6,4 @@ log4j.appender.console.layout.ConversionPattern=%-5p %c{1} - %m%n
 #Loggers
 log4j.rootLogger=ERROR,console
 log4j.logger.com.microsoft=INFO
-log4j.logger.com.microsoft.azure.cosmosdb.kafka=DEBUG
+log4j.logger.com.azure.cosmos.kafka=DEBUG
