+++ src/test/java/com/azure/cosmos/kafka/connect/IntegrationTest.java
@@ -1,0 +1,7 @@
+package com.azure.cosmos.kafka.connect;
+
+/**
+ * To be used with JUnit's {@link org.junit.experimental.categories.Category} to identify integration tests.
+ */
+public interface IntegrationTest {
+}
