--- src/test/java/com/microsoft/azure/cosmosdb/kafka/connect/IntegrationTest.java
@@ -1,7 +1,0 @@
-package com.microsoft.azure.cosmosdb.kafka.connect;
-
-/**
- * To be used with JUnit's {@link org.junit.experimental.categories.Category} to identify integration tests.
- */
-public interface IntegrationTest {
-}
